+++ UXTU4Linux/Assets/Presets/AMDCPU_E.py
@@ -1,8 +1,4 @@
 PRESETS = {
-    "Eco": "--tctl-temp=95",
-    "Balance": "--tctl-temp=95",
-    "Performance": "--tctl-temp=95",
     "Extreme": "--tctl-temp=95",
-    "AC": "--max-performance",
-    "DC": "--power-saving"
+    "AC": "--max-performance"
 }